db: store the values of parameters
Obviously the goal here is to look for cross function bugs.

When we start looking at a function, we search the database to find out how
the function was called.

I've capped it to say that if the function gets called more than 20 times,
we don't know anything about the parameters.  That's sort of arbitrary but
this code is preliminary and will change.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>