db: move buf_size to the new add_definition_db_callback()
This makes the code in smatch_buf_size.c smaller and simpler.
It means we theoretically should start picking up implications between
parameters like:
	some_func(buffer, buf_size);
It also means that instead of getting just the largest buffer size we get
all of them (up to a limit of 20 still).

I had to redo the database.  Everything that modifies the caller_info table
needs to be in the same script so that function_id is calculated
consistently.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>