db: follow calls to functions pointers as well as normal function calls
The Smatch db stuff records anything known about how a function is called.
This is the same thing but for calling function pointers.

The assumption here is that most function pointers are stored in a struct.
So when we store a function pointer in a struct Smatch prints out:

info: sets_fn_ptr '(struct file_operations)->read' 'seq_read'

We store that in the database.

Then when a function is called and something is known about a parameter
we print:
drivers/hid/hid-lg4ff.c +423 lg4ff_init(84)
    info: passes param_value '(struct ff_device)->set_autocenter' 1 '$$' 0

Then after everything is stored in the database, we can query it as we
start parsing the function.  We first find out that we save a pointer to
this function in a struct.  Then we find out how the function pointer is
called.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>