extra, db: record a dummy output for every function call
The smatch_extra.c had a dummy function call to mark calls where nothing
was known.  Smatch extra tried to only print it out where it was needed,
but that gets trickier when you have a bunch of scripts writing to the db.

The new way is to just do it in smatch_db.c.  Do it unconditionally.

I made some related changes to ensure that it would cause a merge_slist()
like it's supposed to.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>