db, extra: introduce add_member_info_callback()
This moves generalizes some infrastructure from smatch_extra.c and moves it
into smatch_db.c so that it can be used by other scripts as well.

Basically when we have a function call we want to print out the states from
any struct members for the structures we pass as parameters.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>