db: limit the number of function pointer calls we check
We already were limiting the number of direct function calls we'd check.
In other words, if the function was called more than 100 times we wouldn't
try to look at all the places where it was called to try figure out what
the parameters were.

This patch does the same thing, except for function pointers.  Somehow I
expected that function pointers would not get called enough where we would
have to cap it at 100.

The problem is that when you try to care about implications it becomes too
much.  Probably eventually we should make a way to merge these things but
throw out all the all the implications.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>