extra: unops don't normally change the value of a variable
The only unops that I can think of which change the value of a variable
are ++ and -- as in x++ means x changed.  The old code said that everything
changed so if you had somthing like:  "if (&a == &b);" then that reset
the values of a and b.  That broke the min() and max() macros.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>