capped: track capped variables across function boundaries
This fixes some false positives.  Imagine the user gives us a struct foo
and we verify that foo.bar is small enough.  We need to track that it was
capped across function calls, otherwise inside the other functions we'll
just know that foo came from the user and not that it was capped.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>