db: pass &foo type parameter information to called functions
If you have a struct declared on the stack, and you pass the address then
the information about that was getting lost before.  But with this patch
passing &foo means that the information about foo.bar gets translated as
$$->bar and passed properly.

In the long term, Smatch needs to move to using memory locations instead of
names like this.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>