db: don't increment the func_id too much
The problem is that there are more scripts now that are printing function
caller info and so that parameter can go down but we're still looking at
the same function.

This still isn't totally right.  A function could be called twice on the
same line.  I suspect smatch_db will need to output more data so we can do
this correctly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>