db: store the return value in the database
If you use the database, then this change makes things a bit slow.  I
haven't totally figured out why.  I should be able to do 100,000 selects
per second on my computer.

But I'm going to push this because it's a step in the right direction and
most people don't build the database.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>