db: handle -1 arguments correctly
Grrr... This is the third time I found that -1 arguments are just being
ignored.  They are supposed to mean that we don't know anything about
the arguments and for a merge with undefined.  But instead what happens
is that you get a NULL slist which when merged with a the defined paths
is dropped.  Just use __unnullify_path() so we have something on the slist
and it forces a merge.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>