extra, ranges: make possibly_true/false() take an expression pointer
The whole "struct dinfo" thing is a pain.  I wanted to use it to handle
bitfields but I haven't gotten around to it.  Passing the expression here
is cleaner.

Also it lets us be more accurate and handle places where we can find an
implied range list, but it's not a variable so there isn't a dinfo set.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>