estate: don't reuse extra_undefined()
Originally I wanted to reuse extra_undefined() as a memory saving measure.
It's common enough to have variables where we don't know what values they
hold so I declared a static smatch_state and reused that for everything.

But the it breaks down when smatch_constraints were introduced.  Internally
Smatch thought all the undefined states were equal to each other so once
we learned something about one then we would assume it was true for all the
others.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>