tree wide: don't print offset into function
This is another formatting change for Artem.  It used to be, that I used
the offset that if the offset into a function changed between two smatch
runs I counted that as a new warning.  It obviously isn't accurate, but
sometimes it's nice to revisit warnings if people are working on the code
around them.

But actually I don't care that much so I'll change it to match what Artem
wants.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>