Makefile: do not install sparse and cgcc
Smatch comes with own version of cgcc and sparse and installs them along with
the 'smatch' binaries. So if the user puts the smatch directory to his PATH,
he may start using smatch's cgcc and sparse version, which may be not what
the user want. Do not install them - the do not seem to be required to be
installed.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>