*new* check_bit_shift.c: (foo & BAR) vs (foo & (1 << BAR))
This makes a list of all the defines which are used as a bit shifter and
complains if they are used in a different place to do a bitwise AND.

The test is not as complete as I would like.  For example, if you have an
enum instead of a define then it doesn't warn about that.  But it did find
four bugs and no false positives so it's better than nothing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>