token_store: creates a position -> token mapping
The problem I had was that I couldn't tell enums from normal values so this
creates a way.  If it's a value and it has a IDENT token tied to it then
either it's a macro or a enum or something.

I used a cursor to track file position so hopefully we don't have to search
through the whole file every time and it's a bit quick.

The problem is that it only tracks the position in the .c file and not in
the included .h files...  Probably I'll fix this eventually.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>