flow: ignore __builtin_constant_p()
The argument to __builtin_constant_p() is evaluated at compile time.  It
doesn't have side effects.

Also we treat it as if it's always false.  known_condition_true/false()
already assume that it's always false.  We call known_condition_true() for
if statements, but we should do it as well for select statements.

For places where we assign a select statement foo = a ? b : c; then we
were using implied_condition_true() which didn't handle
__builtin_constant_p().  But it should have and does now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>