*new* check_macro_side_effects: macros which evaluate the arguments twice
This test didn't find any bugs in the kernel.  People are paranoid about
these but my guess is that they are extremely rare in real life.  I looked
through the git archives but didn't find any.  The closest I found was
people fixing macros to prevent the bug, but actually no one called it with
arguments that had side effects.

Anyway, how this function works is that
1) if you increment or decrement a variable inside a macro.
2) if it has wasn't declared inside the macro.
3) if you incremented or decremented it before.
4) then complain.

There is a long list of ignored macros.  Some of them have deliberate side
effects.  Sometimes they reference variables that weren't passed as
arguments.  There are some like put_user() which have side effects when
you are running Smatch because they have __CHECKER__ defined but when you
compile them for real the side effects are turned off.  (This was done to
add some type checking in Sparse).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>