macro_side_effects: silence false positives inside pre loops
If the macro is inside a pre loop then Smatch evaluates the macro before
the loop and after the loop.  It thinks that the macro parameter is
evaluated twice.

The fix for that is to check if there is a statement outside the macro and
reset the test if there is.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>