db: ignore functions which are called a lot
In the old code, we checked in smatch_db.c how often a function was called
and ignored it if it was called too often.  Instead of doing that, it's
better to not even store the information in the database.  It makes the
database 30% smaller.

The old code sort of ignored functions where the number of times they were
called multiplied by the number of parameters was over 100.  It was ugly
and hacky.  The new way is that if a function is called over 200 times we
ignore it.  The new code ignores fewer functions than the old code (it's
around 700 functions now).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>