extra: add parameter implications
This lets you record parameter implications without writing C.

It's not ideal.  I wanted smatch to know that it's hard to overflow the
count variable in read/write() functions.  This was just a first stab at
that.

A better system would let you parse the code and generate the rules
automatically.  For example if you pass a NULL parameter then it returns
-EINVAL,  There are a lot of simple things like that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>