db: print all return values at end of function
The current code prints the return range whenever we come to a return
statement.  But sometimes we just want to know what the return values are
that the function returns.  That's what this change does.  It combines a
list of all the return ranges and prints out the possible return values
at the end of the function.

It looks like:
fs/ntfs/super.c:541 ntfs_remount() info: function_return_values '(-30),(-22),0'

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>