math: sort of handle conditions
It turns out that "foo = !x;" is handled in a special way in
smatch_condition.c and it doesn't call get_value().  It assumes that foo
can be either 0 or 1 at the end, and does some stuff with implications.

I'm not sure exactly what the correct thing to do is.  This change is a
partial improvement though because now __smatch_implied(1 < 2); produces
what you would expect it to.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>