ranges: commit range_lists_equiv() so that bool_implications compiles...
Grr.  I need to commit this as well or the build breaks.

range_lists_equiv() takes two range lists and returns 1 if they are exactly
the same.  It's used in bool_implications because we record all the states
when we return.  Then we take all the states which are true when we return
1 and all that are true when we return 0.  Then we look at how the states
which concern parameters.  But if the range lists for a parameter are the
same for both 1 and 0 returns then it's not important and we don't
record that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>