sparse: introduce get_expression_value_nomod()
This is an ugly hack, of course.

The problem is that I want to deal with EXPR_SIZEOF expressions but when I
call get_implied_value() on one of these then Sparse turns it into an
EXPR_VALUE behind my back.

I would change it so that Sparse doesn't change sizeof expressions into
value expressions, but Sparse relies on this behavior for some of its
checks.  Probably I should hunt down where that behavior is needed, and
do an explicit call to evaluate_expression() there.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>