flow: print a different filename when --info option is used
The problem here is that static inline functions in the .h file are being
recorded as being in the .h file.  We want them to look like the are in
the .c file.

The output from the --info option is not intended to be human readable
anyway and no one cares if the file is misleading.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>