db: add call_implies table
One way that this can be used is for tracking dereferences across function
boundaries.  You record in the database which function and which argument
have an attribute of type DEREFERENCE.  Then you register a call back which
says "give me all the arguments that have attribute type dereference".

The callbacks are called with the arg expr and a char pointer to the data
in the value column.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>