*new* returns: merge all slists from all returns
This provides a call back:
void all_return_states_hook(void (*callback)(struct state_list *slist));

The way I'm going to use this is to tell if a parameter is derefenced on
every path through a function.  This combines all the end states together
and I'll just do a get_state_slist() to find out which parameters are
always dereferenced.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>