extra: dereferencing a pointer means it is valid
If you see a pointer that is dereferenced, that means it is a valid
pointer.

The check for places that dereference first and then check ignored when
people did unneeded checks and it was clear from the context that the
pointer was non-NULL.  Now that we set the pointer to non-NULL after a
dereference that means we have to turn on those warnings.

It still ends up silencing some warnings, when you check, then dereference,
then check again.  But that used to get two warnings, so now it only gets
one.  No big deal.

The sm_select2.c validation test isn't really doing anything.  I've deleted
it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>