db: move parameter_implications into the database
The intent here was always to move this stuff into the database.  The main
reason to do it now was because calls to global function implications we're
getting over written by static implications.  This stuff is easier to
handle in the database.  That said, I haven't actually got it implemented
yet so only global function implications are stored right now.  I'll fix it
up in a later patch.

I think the new way of storing implications in the database will be more
flexible as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>