buf_size: move get_stored_size() out of get_array_size()
More cleanup.  The stuff in get_stored_size() means we used set_state()
to record the size.  Maybe it was allocated with kmalloc() or something
like that so we know the array size from there.

Also I moved the strip_expr() forward in the get_array_size() function.
That might be a bug fix.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>