buf_size: strcpy() only sets the size of the destination at the end
The problem here is that there are several different ways we calculate the
size of a buffer.  We can look at how it was declared or how maybe we have
it saved as a state.  I moved from preferring how it was declared to using
the state, and it broke the strcpy() checks.  The problem is that we save
the size from strcpy() before check_overflow.c has a chance to see if the
data is larger than the destination.

Really we shouldn't be setting that until after the tests have had a chance
to get the original buffer size.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>