buf_size: remove cast_ratio()
The cast_ratio stuff got broken a few commits back.  It was supposed to
take a expression that had not been passed through strip_expr().  But the
get_array_size() function had a strip_expr() call introduced and it broke
everything.

Just delete this code because I hate it.  Will fix it up shortly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>