buf_size: if the member size is not known then store that in the DB
The thought here is that some drivers allocate a buffer in the
module_init() and it's always the same size.  The problem is that we were
only storing the size if it was known, and not if it was unknown.  So it
created a problem that sometimes we thought we knew the buffer size but
actually we didn't.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>