buf_size: remove the &* and [0] from around the pointer earlier
The situation is that &p[0] is the same as 'p'.  We handled this already
for regular arrays, but not if the array was dynamically allocated.  So
I've moved it forward.

I had to make get_real_array_size() ignore EXPR_BINOP because now it
has to handle &array[5] which is not the same size as &array[0].  It's not
clear if we should return the size of the element or the size of the the
array minus the first 5 elements, but for now we just call it unknown.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>