buf_size: turn off buf size for &addresses unless --spammy is used
This is an ugly temporary hack.  The problem is that people do a lot of
memcpy()ing across multiple struct members so this code, while it's not
wrong, it tends to generate a lot of false positives.

The better way to fix this would be to check in the check_overflow().  Get
the member offset.  Subtract the offset from the full buffer and only warn
if the difference is still too small.  But that code is more complicated to
write.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>