overflow: remove strncpy() false positives
We generate the list of overflow functions automatically.  If we see a
place that does:

	strncpy(foo, bar, sizeof(bar));

then we assume that the second parameter of the function should always be
at least the number of bytes passed in parameter three.  Normally people do
this the right way which is:

	strncpy(foo, bar, sizeof(foo));

But sometimes they do it the wrong way and we need to ignore them.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>