helper: handle if (WARN_ON(i == ARRAY_SIZE(foo)) return -EINVAL
Smatch normally ignores WARN_ON(i == ARRAY_SIZE(foo)); asserts because
normally a condition means something is possible but for those kind of
asserts it means it is impossible.

But in the case of if (WARN_ON()) {... we should treat it like a normal
if statement.  I have introduced a function get_current_statement() that
returns a pointer to the if statement.  We can't just take the current
statement like the current code does because there are several
statements inside the WARN_ON() macro.

Reported-by: Johannes Berg <johannes@sipsolutions.net>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>