Fix a bug with variables named "x86_64".
The code here is equivalent to adding a "#define x86_64 1". But some
kernel code uses variables named "x86_64" and it breaks the build.
Also GCC doesn't handle x86_64 this way.

I pulled this patch from the mailing list, and added a changelog because
it hasn't been applied to Chris's tree yet.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>