deref_check: fix a false positive
The problem is that I thought I didn't need to change the state if it was
merged because that wouldn't trigger an error.  But unfortunately, it can
be merged and then later smatch_implied decides that it &derefed is
implied.  So we need to set the state here and clear out the implications.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>