db: add return_values table
Return values stores all the possible return values for a function.  In
other words it there is one entry per function which is the union of all
the return values.

We already print this information, but the old code used to collect the
return values separately and do a union on that.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>