db: caller_info changes: store the caller, param = -1 is no parameter
If this were a kernel patch I would separate the changes, but Smatch has
lower standards.  :(

First I sometimes want to know which function called me with certain
states.

Next it was using the -1 parameter as an internal value, but I wanted to
use that in my scripts.  So I've introduced an INTERNAL type instead.
Also instead of printing out -1 as the parameter in the --info data, I
print a "call_marker" which is cleaner.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>