db: create return_states table
These patches are sort of broken up in an awkward way.  Sorry for that.

The return_states table is like the caller_info table except instead of
function calls it deals with returns.  You can see from the
fill_db_return_states.pl what data will go there.

I'm going to use the table for a locking check that is not ready to push,
but I may as well reserve the types already.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>