extra, function_hooks, db: change cross function analysis for returns
Before we looked up the return value in return_values and used that to
set the variable in smatch_extra.

What we want ideally is for each return to be an slist.  We merge the
slists together to give the current slist after the function call.  This
would preserve implications.  This patch is a step in that direction.

So basically we store all the information listed in the last paragraph in
the return_states table.  We merge it together.  Done.

The complication is that the return_implies_hook functions do the same
thing and they have to have precedence.  Also the
add_function_assign_hook_extra() conflicts if we use it to set a state
in smatch extra.  And a final conflict is that there is a return_implies
table which does something similar.  This table should be removed
eventually.

I am not ready to push any checks which use the return_states table yet.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>