buf_size, db: if a struct member size is set in the file that's the size
The original Smatch code handled the situation where a buffer which is a
struct member is only allocated once in the entire kernel.

However that situation is pretty rare.  On the other hand, it's pretty
common that you have a struct member which has only one size with in a
particular file.  This patch tries to handle those cases.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>