function_hooks: introduce get_implied_return()
In the old code there was a special hook called
add_function_assign_hook_extra() which was similar to
add_function_assign_hook() except that setting the values in SMATCH_EXTRA
needed to be handled differently.  This patch removes that an instead
introduces a call back called add_implied_return_hook() which takes a
function call and gives the implied range list which is returned.  This
is a better more flexible way of handling things.

This patch doesn't make a lot of difference on it's own, but later I will
add support so that:
	return ERR_PTR(-ENOMEM);
is handled.  In the old code, that wasn't handled because it wasn't an
assignment.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>