extra: delete some debug code
This code was only for debugging and should have been committed in the
first place.

The debugging was to test that we could convert from a range list to
text and then back.  It mostly works.  The problem is that some times
range lists are not connected properly so you get "min-1,2-max" when it
should be "min-max".  It's not harmful to have a range list like that but
it's a bit wrong.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>