math: improve how get_absolute_min/max() work
How it worked before is that get_absolute_max() was the same as
get_implied_max() except that if the value was unknown then it returned
the type_max() instead of whole_range.max.

This is still pretty much the case (for now), but the difference is that if
we do a bitwise AND or a RIGHTSHIFT those are handled differently.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>