sval: create a new basic storage for storing the value of things
Before everything was a long long, but that can't store the value of
ULLONG_MAX.

This patch the sval_t type which holds a ctype as well as the ->value.  It
also adds some helper functions in smatch_sval.c.  Eventually all the
places which currently use "long long" will be updated to use an sval.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>