sval: implement parse_call_math_sval()
This is kind of weird because it's hard to tell which size it should be.
It doesn't really represent a type from the source code, it's just a
value which we calculate ourselves.

I have made it a "long long" type because that should be large enough to
hold the values.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>