sval: start change range_lists to use data_range_sval
I'm trying to transition this in an organized way, in bite sized bits at
a time so I can check my work as I go along.  Range lists are harder
because they are so tied to every subsystem.

Anyway, I introduced the temporary range_list_sval struct and some helper
structs.

The goal was to update smatch_function_hooks.c to using svals.  I've
basically done that but it's a bit crappy.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>