sval: extra: convert match_comparison() to sval
This is sort of complicated and full of temporary hacks.  The problem is
that some of these comparisons should be done as casted to unsigned.  But
because estates are still stored as range_lists instead of range_list_svals
the type information is lost.

So for now I've just cast everything to llong_ctype.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>