sval: convert estates to use range_list_sval internally
This is really the big final push to move all the old code to svals.  There
are still some places which use "long long" so those will have to be fixed
but it's all downhill from here.

I used a bunch of hacks to do it, so those will need to be removed.  Also
there is a lot of deleting old code to be done.  Getting rid of
things like whole_range_list() which don't make sense in the new code.
Fixing how Smatch handles comparisons.  Rename everything back to the old
names.  Bug fixing.  Taking advantage of the new code.  Clean up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>